feat: Add script to download K3d binary
This commit adds a new script, `get-k3d-binary.sh`, which is responsible for downloading the K3d binary from different URLs based on the operating system and architecture. The script uses curl to download the binary and places it in the appropriate directory. The downloaded binaries include `k3d-darwin-amd64`, `k3d-darwin-arm64`, `k3d-linux-amd64`, and `k3d-linux-arm64`.

The purpose of this change is to ensure that the K3d binary used by Devantler.AgeCLI is up to date. This will improve the reliability and performance of the application.

Issue #123